Add mixer type for allocated actuator
The mixer is added with the following character string in a *.mix file:
`A: X`
with X a number between 0 and 15
This indicates that the actuator number X will be written to the output.

Ideally, this mixer would use the topic `vehicle_actuator_setpoint`, but for now it receives info on topics `actuator_controls_4` and `actuator_controls_5`.